# api/openapi.yaml
openapi: 3.1.0
info:
  title: Loom CLI API
  description: |
    REST API for the Loom CLI web UI and daemon. Covers workspace management,
    issue tracking, agent lifecycle, terminal sessions, git operations, fleet
    coordination, monitoring, and observability.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server (post-consolidation single server)
...
security:
  - BearerAuth: []
paths:
...
  /api/workspaces/{ws}/agents:
    get:
      operationId: listAgents
      tags: [agent-control]
      summary: List agents via daemon control socket
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentControlEntry"
        "503":
          description: Daemon unavailable
    post:
      operationId: createAgent
      tags: [agent-control]
      summary: Create an agent assignment
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role_name]
              properties:
                workspace_key:
                  type: string
                name:
                  type: string
                role_name:
                  type: string
                kind:
                  type: string
                  description: Optional role kind for creating an interactive role.
                prompt:
                  type: string
                  description: Literal inline prompt text for interactive roles.
                prompt_file:
                  type: string
                  description: Custom or builtin prompt selector for interactive roles.
                auto:
                  type: boolean
                backend:
                  type: string
                fallback_backends:
                  type: array
                  items:
                    type: string
                runtime_provider:
                  type: string
                  enum: [local, e2b, kubernetes, daytona, ci, other]
                repos:
                  type: array
                  items:
                    type: string
                repo_groups:
                  type: array
                  items:
                    type: string
                cross_repo:
                  type: boolean
                parent:
                  type: string
                desired_state:
                  type: string
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAgentInfo"
        "400":
          description: Invalid agent create request
        "409":
          description: Agent already exists
        "503":
          description: Agent service unavailable
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. The key is a 64-character hex string generated
        from 32 random bytes. Configured via `LOOM_AUTH_API_KEY` env var or
        loaded from `~/.loom/api-key`. Supports fallback to `?token=` query
        parameter for WebSocket/EventSource connections.

    WorkerToken:
      type: http
      scheme: bearer
      description: |
        Static token for internal worker API endpoints. Configured via
        `LOOM_WORKER_TOKEN` env var. If not set, worker API returns 503.

    FleetApiKey:
      type: apiKey
      in: header
      name: X-Fleet-API-Key
      description: |
        API key for fleet registration. Configured via `LOOM_FLEET_API_KEY`
        env var. Constant-time comparison with optional per-IP rate limiting.

    FleetJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for fleet claim/heartbeat endpoints. HMAC-SHA256 signed.
        Configured via `LOOM_FLEET_JWT_KEY` env var. Token expiry: 1 hour.
  parameters:
    WorkspaceId:
      name: ws
      in: path
      required: true
      description: Workspace identifier
      schema:
        type: string
...
  schemas:
...
    WorkspaceAgentInfo:
      type: object
      required: [name, repos, repo_groups, cross_repo]
      properties:
        name:
          type: string
        repos:
          type: array
          items:
            type: string
          default: []
        repo_groups:
          type: array
          items:
            type: string
          default: []
        cross_repo:
          type: boolean
        last_provision_outcome:
          type: string
          enum: [in_progress, succeeded, failed]
          description: Latest eager Daytona lead provision-attempt outcome.
        last_provision_error:
          type: string
          description: Bounded error from the latest failed provision attempt.
        last_provision_at:
          type: string
          format: date-time
          description: Time the latest provision-attempt outcome was recorded.
...
    AgentControlEntry:
      type: object
      description: Agent entry from daemon control socket list
      required: [name, role, status]
      properties:
        name:
          type: string
        role:
          type: string
        status:
          type: string

    # ── Session Schemas ─────────────────────────────────────────────────
...